make zn_server struct public
