+++ zen/server.c
@@ -2,51 +2,13 @@
 
 #include <stdio.h>
 #include <wayland-server.h>
-#include <wlr/backend.h>
-#include <wlr/render/allocator.h>
-#include <wlr/render/wlr_renderer.h>
-#include <wlr/types/wlr_compositor.h>
-#include <wlr/types/wlr_output.h>
-#include <wlr/types/wlr_xdg_shell.h>
-#include <wlr/xwayland.h>
 #include <zen-desktop-protocol.h>
 
 #include "zen-common/log.h"
 #include "zen-common/util.h"
-#include "zen/display-system.h"
-#include "zen/input-manager.h"
 #include "zen/output.h"
-#include "zen/scene/scene.h"
 #include "zen/xdg-toplevel-view.h"
 #include "zen/xwayland-view.h"
-
-struct zn_server {
-  struct wl_display *display;
-  struct wl_event_loop *loop;
-  struct wlr_backend *backend;
-  struct wlr_renderer *renderer;
-  struct wlr_allocator *allocator;
-  struct wlr_xwayland *xwayland;
-
-  // these objects will be automatically destroyed when wl_display is destroyed
-  struct wlr_compositor *w_compositor;
-  struct wlr_xdg_shell *xdg_shell;
-
-  struct zn_display_system *display_system;
-  struct zn_input_manager *input_manager;
-
-  struct zn_scene *scene;
-
-  char *socket;
-
-  struct wl_listener new_input_listener;
-  struct wl_listener new_output_listener;
-  struct wl_listener xdg_shell_new_surface_listener;
-  struct wl_listener display_system_switch_listener;
-  struct wl_listener xwayland_new_surface_listener;
-
-  int exit_code;
-};
 
 static void
 zn_server_new_input_handler(struct wl_listener *listener, void *data)
@@ -131,24 +93,6 @@ zn_server_xwayland_new_surface_handler(struct wl_listener *listener, void *data)
   wlr_xwayland_surface_ping(xwayland_surface);
 
   (void)zn_xwayland_view_create(xwayland_surface, self);
-}
-
-struct wl_event_loop *
-zn_server_get_loop(struct zn_server *self)
-{
-  return self->loop;
-}
-
-struct wlr_renderer *
-zn_server_get_renderer(struct zn_server *self)
-{
-  return self->renderer;
-}
-
-struct zn_scene *
-zn_server_get_scene(struct zn_server *self)
-{
-  return self->scene;
 }
 
 int
