+++ zen/xdg-toplevel-view.c
@@ -21,7 +21,7 @@ zn_xdg_toplevel_view_map(struct wl_listener* listener, void* data)
 {
   struct zn_xdg_toplevel_view* self =
       zn_container_of(listener, self, map_listener);
-  struct zn_scene* scene = zn_server_get_scene(self->server);
+  struct zn_scene* scene = self->server->scene;
   struct zn_screen* screen;
   UNUSED(data);
